primeiro funcional
