+++ Lexical/definicoes.h
