+++ Lexical/funcs_iniciais.c
@@ -1,387 +1,17 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <ctype.h>
-#include "definicoes.h"
+#include "funcs_iniciais_lib.h"
 
-
-typedef struct Node Node;
-
-struct Node{ 
-  token tk;          // ver ponteiro / string 
-  Node *prev;
-}; 
-
-char caractere;
-char *filename; 
 FILE *fp; 
-Node *atual;
-int linha;
-
-void ler();
-void printList();
-void trata_digito(token *tk);
-void trata_atribuicao(token *tk);
-void trata_aritmetico(token *tk);
-void trata_pontuacao(token *tk);
-void insere_lista(token *tk);
-void trata_ident_reserv(token *tk);
-void trata_relacional(token *tk);
-void print_token(token *tk);
-void printList();
-void desalocador();
-void pega_token(token *tk);
-void AnalisadorLexicalN1();
-void lexical_error(int n, token *tk);
-
-void ler()
-{ 
-    caractere = fgetc(fp); 
-} 
-/// Cria um token baseado num array de char passado
-void set_token_s(token *tk, char string[])
-{ 
-    memset (tk->lexema, 0, lexema_size_max);
-    memcpy ( tk->lexema, string, strlen(string)+1 ); 
-} 
-
-/// Cria um token baseado no caractere atual
-void set_token_c(token *tk)
-{ 
-    memset (tk->lexema, 0, lexema_size_max);
-    memcpy ( tk->lexema, &caractere, sizeof(caractere) ); 
-} 
-
-void trata_digito(token *tk){ 
-    char num[lexema_size_max]; 
-    while(isdigit(caractere)){ 
-        strncat(num, &caractere, 1);  
-        ler(); 
-    }
-    set_token_s (tk, num);
-    tk->simbolo = snúmero;
-} 
-
-void trata_atribuicao(token *tk){ 
-    ler(); 
-    if(caractere == '='){ 
-        set_token_s (tk, ":="); 
-        tk->simbolo = satribuicao; 
-        ler(); 
-    }else{  
-        set_token_s (tk, ":"); 
-        tk->simbolo = sdoispontos; 
-    }
-}
-
-void trata_aritmetico(token *tk){ 
-    set_token_c (tk);  
-    switch(caractere){
-        case '+': 
-            tk->simbolo = smais;
-            break;
-        case '-': 
-            tk->simbolo = smenos ;
-            break;
-        case '*':  
-            tk->simbolo = smult ;
-            break;
-    }
-    ler(); 
-} 
-
-void trata_pontuacao(token *tk){
-    set_token_c (tk);  
-    switch (caractere)
-    {
-    case ',':
-        tk->simbolo = svírgula;
-        break;
-    case '(':
-        tk->simbolo = sabre_parenteses;
-        break;
-    case ')':
-        tk->simbolo = sfecha_parenteses;
-        break;
-    case ';':
-        tk->simbolo = sponto_virgula;
-        break;    
-    case '.':
-        tk->simbolo = sponto;
-        break;    
-    default:
-        break;
-    }
-    ler();
-} 
-
-void insere_lista(token *tk)
-{
-    Node *no = (Node*) malloc(sizeof(Node));
-    memcpy ( &(no->tk), tk, sizeof(token) );
-
-    if(atual != NULL)
-    {
-        no->prev = atual;
-    }    
-    else
-    {
-        no->prev = NULL;
-    }
-
-    atual = no;
-}
-
-void trata_ident_reserv(token *tk){
-    char id[lexema_size_max] = "";
-    strncat(id, &caractere, 1);  
-    ler();
-    
-    while(isalnum(caractere) || (caractere == 95)){
-        strncat(id, &caractere, 1); 
-        ler();
-    }
-    char ao = caractere;
-    set_token_s(tk, id);
-
-    if (strcmp(id, "programa") == 0) {
-        tk->simbolo = sprograma;
-    } else if (strcmp(id, "se") == 0) {
-        tk->simbolo = sse;
-    }else if (strcmp(id, "entao") == 0) {
-        tk->simbolo = sentao;
-    }else if (strcmp(id, "senao") == 0) {
-        tk->simbolo = ssenao;
-    }else if (strcmp(id, "enquanto") == 0) {
-        tk->simbolo = senquanto;
-    }else if (strcmp(id, "faca") == 0) {
-        tk->simbolo = sfaca;
-    }else if (strcmp(id, "inicio") == 0) {
-        tk->simbolo = sinício;
-    }else if (strcmp(id, "fim") == 0) {
-        tk->simbolo = sfim;
-    }else if (strcmp(id, "escreva") == 0) {
-        tk->simbolo = sescreva;
-    }else if (strcmp(id, "leia") == 0) {
-        tk->simbolo = sleia;
-    }else if (strcmp(id, "var") == 0) {
-        tk->simbolo = svar;
-    }else if (strcmp(id, "inteiro") == 0) {
-        tk->simbolo = sinteiro;
-    }else if (strcmp(id, "booleano") == 0) {
-        tk->simbolo = sbooleano;
-    }else if (strcmp(id, "verdadeiro") == 0) {
-        tk->simbolo = sverdadeiro;
-    }else if (strcmp(id, "falso") == 0) {
-        tk->simbolo = sfalso;
-    }else if (strcmp(id, "procedimento") == 0) {
-        tk->simbolo = sprocedimento;
-    }else if (strcmp(id, "funcao") == 0) {
-        tk->simbolo = sfuncao;
-    }else if (strcmp(id, "div") == 0) {
-        tk->simbolo = sdiv;
-    }else if (strcmp(id, "e") == 0) {
-        tk->simbolo = se;
-    }else if (strcmp(id, "ou") == 0) {
-        tk->simbolo = sou;
-    }else if (strcmp(id, "nao") == 0) {
-        tk->simbolo = snao;
-    }else{
-        tk-> simbolo = sidentificador;
-    }
-} 
-
-void trata_relacional(token *tk){
-    char switchcaractere = caractere;
-    ler();
-    switch(switchcaractere){
-        case '!':
-            if(caractere == '='){
-                set_token_s(tk, "!=");
-                tk->simbolo = sdif;
-                ler();
-            }
-            else{
-                ungetc(caractere, fp);
-                lexical_error(4, tk);
-            }
-            break;
-        case '<':
-            if(caractere == '='){
-                set_token_s(tk, "<=");
-                tk->simbolo = smenorig;
-                ler();
-            }
-            else{
-                set_token_s(tk, "<");
-                tk->simbolo = smenor;
-            }
-            break;
-        case '>':
-            if(caractere == '='){
-                set_token_s(tk, ">=");
-                tk->simbolo = smaiorig;
-                ler();
-            }
-            else{
-                set_token_s(tk, ">");
-                tk->simbolo = smaior;
-            }
-            break;
-        case '=':
-            set_token_s(tk, "=");
-            tk->simbolo = sig;
-            break;
-    }
-}
-
-void print_token(token *tk){
-    printf("Lexema: %s\tSimbolo: %s\n", tk->lexema, print_enum(tk->simbolo));
-    printf("==================================================================================\n");
-}
-
-void printList()
-{
-    Node *no = atual;
-    printf("==================================================================================\n");
-    while (no->prev != NULL)
-    {
-        print_token(&(no->tk));
-        no = no->prev;
-    }
-    print_token(&(no->tk));
-}
-
-void desalocador()
-{
-    Node *no = atual;
-    while (no->prev != NULL)
-    {
-        atual = no->prev;
-        free(no);
-        no = atual;
-    }    
-    free(no);
-}
-
-void pega_token(token *tk){ 
-   char aux = caractere;
-    if(isdigit(caractere)){ 
-        trata_digito(tk);  
-    }else{ 
-        if(isalpha(caractere)){ 
-            trata_ident_reserv(tk); 
-        }else{ 
-            if(caractere == ':'){ 
-                trata_atribuicao(tk); 
-            }else{ 
-                if(caractere == '+' || caractere == '-' || caractere == '*'){ 
-                    trata_aritmetico(tk); 
-                }else{ 
-                    if(caractere == '!' || caractere == '<' || caractere == '>' || caractere == '='){ 
-                        trata_relacional(tk); 
-                    }else{ 
-                        if(caractere == ';' || caractere == ','|| caractere == '('|| caractere == ')'|| caractere == '.'){ 
-                            trata_pontuacao(tk); 
-                        }else{ 
-                            lexical_error(3, tk); 
-                        } 
-                    } 
-                } 
-            } 
-        } 
-    }
-} 
-
-void lexical_error(int n, token *tk){
-    char* erros[] = {"",
-        "1: Chaves Abertas ('{') sem chaves fechadas depois (era esperado '}')",
-        "2: Chaves Fechadas ('}') sem chaves abertas antes (era esperado '{' antes)",
-        "3: Caractere Inválido",
-        "4: Ocorrencia de Exclamação ('!') sem Sinal de Igual ('=') depois (símbolo inválido)" 
-    };
-    
-    printf("Erro lexical");
-    printf(" %s ", erros[n]);
-    printf("na linha %d\n", linha);
-
-    
-    set_token_s(tk, "Erro");
-    tk->simbolo = serro;
-
-    ler();
-}
-
-void AnalisadorLexicalN1()
-{
-    filename = "teste_3.txt";
-    fp = fopen(filename, "r");
-
-    if (fp == NULL) 
-    { 
-        printf("Error: could not open file %s", filename); 
-        return; 
-    } 
-
-    linha = 1;
-
-    ler(); 
-
-    while (caractere != EOF)
-    {
-        while ((caractere == '{' || caractere == ' ' || caractere == '\n' || caractere == '}') && caractere != EOF) 
-        {        
-            if(caractere == '{') 
-            {
-                int abre = 0;
-                    while((caractere != '}' || abre > 1) && caractere != EOF) 
-                    { 
-                        if(caractere == '{') 
-                            abre++;
-                        if(caractere == '}') 
-                            abre--;
-                        if(caractere == '\n')
-                            linha++;
-                        ler(); 
-                    }            
-                    if(caractere == EOF) 
-                    { 
-                        token tk;
-                        lexical_error(1, &tk);
-                        insere_lista(&tk);
-                    } 
-                    ler(); 
-            } 
-
-            if(caractere == '}'){
-                token tk;
-                lexical_error(2, &tk);
-                insere_lista(&tk);  
-            }
-
-            while(caractere == ' ' || caractere == '\n')
-            {
-                if(caractere == '\n'){
-                    linha++;
-                }
-                ler();
-            }
-        }             
-        
-        if(caractere != EOF) 
-        { 
-            ///printf("%c", caractere);
-            token tk;
-            pega_token(&tk);
-            insere_lista(&tk);
-        }
-    }
-    
-    printList();
-    printf("linhas: %d", linha);
-    desalocador();
-    fclose(fp); 
-}
+char* filename = "new.txt";
+token tk;
+int linha_main = 0;
 
 int main(){
-    AnalisadorLexicalN1();
+    fp = fopen(filename, "r");
+    caractere = fgetc(fp); 
+    while ( feof(fp) != 1)
+    { 
+        AnalisadorLexical(fp,&linha_main,&tk);
+        print_token(&tk);
+    }
+    print_linhas();
 }
